+++ src/test/fixtures/taglib/DemoUITagLib.groovy
@@ -1,0 +1,20 @@
+class DemoUITagLib {
+    static namespace = "demoUI"
+
+    def accountLink = { attrs, body ->
+        List<String> linkAttributes = ["href", "class", "target"]
+        out << attrs.href
+        out << attrs['class']
+        if (attrs.containsKey('isAtlas')) { }
+    }
+
+    def adminButton = { attrs, body ->
+        out << body()
+        out << attrs.url
+        out << attrs.class
+    }
+
+    def messagePrinter = { attrs ->
+        out << attrs.isAtlas
+    }
+}